Change the format for the market cap and volume numbers. Change the default currency after pressing the search button. Change the synchronization of the UIPicker with the SearchButton. Change the font and the colour of the UIPicker labels. Change the volume value and the market cap value into integers (for rounding up).

+++ CryptoTracker/Model/CryptoDataModel.swift
@@ -13,8 +13,8 @@ class CryptoDataModel{
     // properties of the model
     var cryptoName : String = ""
     var cryptoPrice : Double = 0.0
-    var marketCap : Double = 0.0
-    var volume: Double = 0.0
+    var marketCap : Int = 0
+    var volume: Int = 0
     var changePercentage : Double = 0.0
     
 }
